requirements.txt
boto3>=1.28
pandas>=2.0
streamlit>=1.36
sqlalchemy>=2.0
psycopg2-binary>=2.9
botocore>=1.31
toml
python-dotenv

# API dependencies (optional - only needed if API endpoints are enabled)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
